requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
sqlalchemy==2.0.35
pymysql==1.1.1
pandas==2.2.2
openpyxl==3.1.5
python-dotenv==1.0.1
langgraph==0.2.0
langchain==0.3.0
langchain-community==0.3.0
ollama==0.3.3
pydantic==2.9.2
python-jose==3.3.0
passlib==1.7.4
httpx==0.27.0